// frontend/src/app/models/master/masSharePercentHistoryData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface MasSharePercentHistoryData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    /** ส่วนแบ่งคณะ */
    costShareDepPercent?: number;
    /** ส่วนแบ่งมหาลัย */
    costShareGlobalPercent?: number;
    /** ส่วนแบ่งศูนย์บริการวิชาการ*/
    costShareCenterPercent?: number;
    /** สร้างโดย */
    createByUserId?: number;
    /** วันที่บันทึก */
    createDate?: string;
    /** fk (dep_id) */
    depId?: number;
    first?: number;
    /** ผู้บันทึก */
    firstName?: string;
    mode?: string;
    rowNum?: number;
    /** PK auto run (share_percent_history_id)  */
    sharePercentHistoryId?: number;
    size?: number;

    /** ผู้บันทึก (ภาษาไทย) */
    fullNameTh?: string;
    /** ผู้บันทึก (ภาษาอังกฤษ) */
    fullNameEn?: string;
}